requirements.txt
uvicorn
gpt4all
python-dotenv
fastapi